incorporate sweep to scan lidar fix into LidarIntegrator.py
