+++ mapping/pipeline/ldr_to_h5.py
@@ -9,6 +9,7 @@ from GPSTransforms import IMUTransforms
 from LidarTransforms import loadLDR, loadLDRCamMap
 from pipeline_config import EXPORT_STEP, EXPORT_START, EXPORT_NUM, LANE_FILTER,\
     PARAMS_TO_LOAD, OPT_POS_FILE, EXPORT_FULL_NUM_FILE
+from LidarIntegrator import transform_points_in_sweep
 
 '''
 Essentially just pieces from LidarIntegrator except avoids
@@ -16,35 +17,6 @@ storing the data for all time steps in memory
 
 Writes full point clouds for scan matching later
 '''
-
-def interp_transforms(T1, T2, alpha):
-    assert alpha <= 1
-    T_avg = alpha * T1 + (1 - alpha) * T2
-    ## Need to orthonormalize transform
-    R = T_avg[0:3, 0:3]
-    R = np.linalg.qr(R, mode='complete')[0]
-    T_avg[0:3, 0:3] = R
-    return T_avg
-
-
-def transform_points_in_sweep(pts, times, fnum, imu_transforms):
-    for time in set(times):
-        mask = times == time
-
-        # FIXME PARAM
-        offset = (time / float(1e6)) / 0.02
-        #assert offset <= 5.1, offset   # FIXME HACK
-        offset = min(5, offset)
-
-        ind1 = int(fnum - math.ceil(offset))
-        ind2 = int(fnum - math.floor(offset))
-        # FIXME Can't interpolate before 0 so first frame not shifted properly
-        ind1, ind2 = max(ind1, 0), max(ind2, 0)
-        transform = interp_transforms(imu_transforms[ind1, :, :], imu_transforms[ind2, :, :], offset / 5.0)
-
-        # transform data into imu_0 frame
-        pts[:, mask] = np.dot(transform, pts[:, mask])
-
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Convert ldr files to h5 files containing points')
